+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Add `-L / --dereference` option.
 - Add `-H / --dereference-command-line` and `--dereference-command-line-symlink-to-dir` options.
 - Add `--full-time` option.
+- Add `--group-directories-first`.
 
 ### Changed